requirements.txt
aiogram
httpx
aiocqhttp
tgs
imageio
janus
filetype
cairosvg
Pillow
coloredlogs
ffmpy
line-bot-sdk
quart